% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCohort.R
\name{computeSccStats}
\alias{computeSccStats}
\title{Get Zipped or manifest Scc stat Results}
\usage{
computeSccStats(
  connection,
  config,
  analysisIds = NULL,
  studyName = NULL,
  outcomeCohortIds = getDefaultOutcomeIds(connection, config),
  targetCohortIds = getDefaultExposureIds(connection, config)
)
}
\arguments{
\item{connection}{DatabaseConnector connection}

\item{config}{cdm config loaded with loadCdmConfig function}

\item{studyName}{- optional string to name this study for result management}

\item{outcomeCohortIds}{- vector of outcome cohort ids or NULL}

\item{targetCohortIds}{- vector of exposure cohort ids or NULL}
}
\description{
Get zip files for scc
Partial reward execution with a subset of targets or outcomes. If both are null this will generate SCC results for all
exposure and outcome pairs. This is only really useful if you're adding an cohort after the full result set has been
generated.
}
