% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{createCdmConfiguration}
\alias{createCdmConfiguration}
\title{Create CDM configuration file}
\usage{
createCdmConfiguration(
  cdmConfigPath,
  keyring = NULL,
  overwrite = FALSE,
  testConnection = TRUE
)
}
\arguments{
\item{cdmConfigPath}{path to cdm configuration file}

\item{keyring}{keyring::keyring. For systems that support multiple keyrings, specify the name of the keyring to use here.
If NULL, then the default keyring is used.}

\item{overwrite}{Overwite existing file (if it exists)}

\item{testConnection}{Attempt to connect to database and write to schemas needed for writing?}
}
\description{
Opens a file for editing that contains the default settings for a cdm
}
