% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Execute.R
\name{createResultsManifest}
\alias{createResultsManifest}
\title{Create Results Manifest (s3)}
\usage{
createResultsManifest(config, studyName = NULL)
}
\arguments{
\item{config}{Reward config object}
}
\description{
Create json manifest for results files to upload
}
