% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Execute.R
\name{createResultsZip}
\alias{createResultsZip}
\title{Create Results Zip}
\usage{
createResultsZip(config)
}
\arguments{
\item{config}{Reward config object}
}
\description{
Create zip for results including CDM meta-data
}
