% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCohort.R
\name{createSccRiskWindowTable}
\alias{createSccRiskWindowTable}
\title{Get Self controlled cohort risk windows}
\usage{
createSccRiskWindowTable(
  connection,
  config,
  analysisSettings,
  exposureIds = getDefaultExposureIds(connection, config),
  analysisId = NULL
)
}
\arguments{
\item{connection}{DatabaseConnector connection}

\item{config}{cdm configuration}

\item{analysisSettings}{Id of analysis}

\item{exposureIds}{Outcomes to run for}

\item{analysisId}{cdm configuration}
}
\description{
Get Self controlled cohort risk windows
}
