% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{createTargetsFile}
\alias{createTargetsFile}
\title{Create targets file}
\usage{
createTargetsFile(path = "_targets.R", overwrite = FALSE)
}
\arguments{
\item{path}{Path where targets file will live - targets expects _targets.R by default}
}
\description{
Targets is a workflow manager for R execution tasks that prevents the reptition of work.
This utility function creates a targets file that can be used to execute Reward on a CDM
}
