% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportResults.R
\name{exportResults}
\alias{exportResults}
\title{Export results in to a single zipped file}
\usage{
exportResults(
  config,
  exportZipFile = paste0("reward-results-", config$database, ".zip")
)
}
\arguments{
\item{config}{CdmConfiguration object}

\item{exportZipFile}{Name of zip file to export}
}
\description{
Export results in to a single zipped file
}
\details{
Zips files and records their checksums
}
