% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCohort.R
\name{exportSccTarStats}
\alias{exportSccTarStats}
\title{Clean up and export time at risk stats}
\usage{
exportSccTarStats(tarStats, config, analysisId)
}
\arguments{
\item{tarStats}{object returned by SelfControlledCohort getSccRiskWindowStats}

\item{config}{cdm configuration  object}

\item{analysisId}{id to mark output files with}

\item{statType}{string - one of "treatmentTimeDistribution", "timeToOutcomeDistribution","timeToOutcomeDistributionExposed",
"timeToOutcomeDistributionUnexposed"}
}
\description{
Prepares CSV ouput of time at risk stats for uploading to database
}
