% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructCohorts.R
\name{generateAtlasCohortSet}
\alias{generateAtlasCohortSet}
\title{Generate bulk cohorts with cohort generator}
\usage{
generateAtlasCohortSet(config, connection = NULL)
}
\arguments{
\item{config}{CdmConfig object}

\item{connection}{DatabaseConnector::connection instance}
}
\description{
Use cohort generator to create cohorts
}
