% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructCohorts.R
\name{getAtlasCohortDefinitionSet}
\alias{getAtlasCohortDefinitionSet}
\title{Get Atlas Cohort Definition Set}
\usage{
getAtlasCohortDefinitionSet(config)
}
\arguments{
\item{config}{CdmConfig object}
}
\description{
Get Reward atlas cohort definitions
}
