% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCohort.R
\name{getDefaultExposureIds}
\alias{getDefaultExposureIds}
\title{get exposure cohort ids}
\usage{
getDefaultExposureIds(connection, config)
}
\arguments{
\item{connection}{DatabaseConnector connection}

\item{config}{cdm configuration}
}
\description{
Return all outcome cohort ids
}
