% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCohort.R
\name{getDefaultOutcomeIds}
\alias{getDefaultOutcomeIds}
\title{get outcome cohort ids}
\usage{
getDefaultOutcomeIds(connection, config)
}
\arguments{
\item{connection}{DatabaseConnector connection}

\item{config}{cdm configuration}
}
\description{
Return all outcome cohort ids
}
