% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCohort.R
\name{getDefaultSccDataFileName}
\alias{getDefaultSccDataFileName}
\title{Get default results file name}
\usage{
getDefaultSccDataFileName(config, analysisId, position = NULL)
}
\description{
Get default results file name
}
