% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCohort.R
\name{getSccSettings}
\alias{getSccSettings}
\title{Get scc settings stored in database}
\usage{
getSccSettings(connection, config, analysisIds = NULL)
}
\description{
returns a set of json objects that are SCC parameters
}
