% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportReferenceFiles.R
\name{importReferenceTables}
\alias{importReferenceTables}
\title{Import reference tables}
\usage{
importReferenceTables(
  connection,
  cdmConfig,
  zipFilePath,
  overwriteReferences = FALSE
)
}
\arguments{
\item{connection}{DatabaseConnector connection}

\item{cdmConfig}{cdmConfig object}

\item{zipFilePath}{zip file path}
}
\description{
Note that this always overwrites the existing reference tables stored in the database
}
