% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{loadCdmConfiguration}
\alias{loadCdmConfiguration}
\title{load cdm config object}
\usage{
loadCdmConfiguration(cdmConfigPath, keyring = NULL)
}
\arguments{
\item{cdmConfigPath}{path to cdm configuration file}

\item{keyring}{keyring::keyring. For systems that support multiple keyrings, specify the name of the keyring to use here.
If NULL, then the default keyring is used.}
}
\description{
Loads config and prompt user for db password
Password can be set in envrionment variable passwordEnvironmentVariable of yaml file
}
