% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RewardExecutionPackage.R
\docType{package}
\name{RewardExecutionPackage-package}
\alias{RewardExecutionPackage-package}
\title{RewardExecutionPackage: REWARD Execution Package}
\description{
Study execution package for generating REWARD results. Uses SelfControlledCohort package to generate effect estimates. Handles uploading of results to reward system.
}
\keyword{internal}
