% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCohort.R
\name{runScc}
\alias{runScc}
\title{Peform SCC from self controlled cohort package with rewardbs settings}
\usage{
runScc(
  connection,
  config,
  analysisSettings,
  analysisId,
  postProcessFunction = batchStoreSccResults,
  postProcessArgs = list(config = config, analysisId = analysisId),
  exposureIds = getDefaultExposureIds(connection, config),
  outcomeIds = getDefaultOutcomeIds(connection, config),
  cores = parallel::detectCores() - 1
)
}
\description{
Used by targets for execution of steps. See computeSccResults for other settings
}
