% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{validateCdmConfigFile}
\alias{validateCdmConfigFile}
\title{Validate a cdm configuration file}
\usage{
validateCdmConfigFile(cdmConfigPath, testConnection = TRUE, keyring = NULL)
}
\arguments{
\item{cdmConfigPath}{path to cdm configuration file}

\item{testConnection}{Attempt to connect to database and write to schemas needed for writing?}

\item{keyring}{keyring::keyring. For systems that support multiple keyrings, specify the name of the keyring to use here.
If NULL, then the default keyring is used.}
}
\description{
Opens a file for editing that contains the default settings for a cdm
}
