% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructCohorts.R
\name{computeAtlasCohorts}
\alias{computeAtlasCohorts}
\title{Compute Atlas cohorts}
\usage{
computeAtlasCohorts(connection, config, exposureCohorts = FALSE)
}
\arguments{
\item{connection}{DatabaseConnector connection to cdm}

\item{config}{cdmConfiguration object}

\item{exposureCohorts}{Exposure or outcome based cohorts?}
}
\description{
Computes sql cohorts against the CDM
}
