% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Execute.R
\name{execute}
\alias{execute}
\title{Execute package}
\usage{
execute(
  cdmConfigPath,
  referenceZipFile,
  deleteExistingCohorts = FALSE,
  analysisIds = NULL
)
}
\arguments{
\item{cdmConfigPath}{path to cdm configuration file}

\item{referenceZipFile}{Path to rewardb cohort references zip file}

\item{deleteExistingCohorts}{If results for cohorts already exist, delete and start again?}

\item{analysisIds}{optionally only run certain analysis types}
}
\description{
Upload cohorts and references
Execute SCC methods
Store results (and optionally transfer them to main REWARD server location)
}
