% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCohort.R
\name{getSccRiskWindowStats}
\alias{getSccRiskWindowStats}
\title{Get Self controlled cohort risk windows}
\usage{
getSccRiskWindowStats(
  connection,
  config,
  analysisSettings,
  outcomeIds = getDefaultOutcomeIds(connection, config),
  analysisId = NULL
)
}
\arguments{
\item{connection}{DatabaseConnector connection}

\item{config}{cdm configuration}

\item{analysisSettings}{Id of analysis}

\item{outcomeIds}{Exposure to run for}

\item{analysisId}{cdm configuration}
}
\description{
Get Self controlled cohort risk windows
}
