% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCohort.R
\name{runScc}
\alias{runScc}
\title{Peform SCC from self controlled cohort package with rewardbs settings}
\usage{
runScc(
  connection,
  config,
  postProcessFunction,
  postProcessArgs,
  analysisSettings,
  analysisId,
  exposureIds,
  outcomeIds,
  cores = parallel::detectCores() - 1
)
}
\description{
Peform SCC from self controlled cohort package with rewardbs settings
}
