% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{validateCdmConfigFile}
\alias{validateCdmConfigFile}
\title{Validate a cdm configuration file}
\usage{
validateCdmConfigFile(
  cdmConfigPath,
  testConnection = TRUE,
  keyring = NULL,
  testS3Rw = TRUE
)
}
\arguments{
\item{cdmConfigPath}{path to cdm configuration file}

\item{testConnection}{Attempt to connect to database and write to schemas needed for writing?}

\item{keyring}{keyring::keyring. For systems that support multiple keyrings, specify the name of the keyring to use here.
If NULL, then the default keyring is used.}

\item{testS3Rw}{If useAwsS3Export is set to true, test reding and writing of objects to s3}
}
\description{
Opens a file for editing that contains the default settings for a cdm
}
